add roslint and fix linter errors
