+++ mission_manager/include/mission_manager/behaviors/fixed_rudder.h
@@ -34,42 +34,63 @@
 
 // Original version: Christopher Scianna Christopher.Scianna@us.QinetiQ.com
 
-#ifndef __ALTITUDE_HEADING_H
-#define __ALTITUDE_HEADING_H
+/* Behavioral fixed rudder
 
-#include "../behavior.h"
-#include "mission_manager/AltitudeHeading.h"
+        <fixed_rudder>
+            <description>
+                00:00:00 - .
+            </description>
+            <when unit="sec">0</when>
+            <timeout unit="sec">50</timeout>
+            <depth unit="deg">10.0</depth>
+            <rudder unit="deg">180.0</rudder>
+            <speed_knots>0.0</speed_knots>
+        </fixed_rudder>
+*/
+
+#ifndef MISSION_MANAGER_BEHAVIORS_FIXED_RUDDER_H
+#define MISSION_MANAGER_BEHAVIORS_FIXED_RUDDER_H
+
+#include <string>
+#include "mission_manager/behavior.h"
+#include "mission_manager/FixedRudder.h"
 
 namespace mission_manager
 {
-class AltitudeHeadingBehavior : public Behavior
+
+class FixedRudderBehavior : public Behavior
 {
  public:
-  AltitudeHeadingBehavior();
-  virtual ~AltitudeHeadingBehavior();
+  FixedRudderBehavior();
+  virtual ~FixedRudderBehavior();
   virtual bool parseMissionFileParams();
-  bool getParams(ros::NodeHandle nh);
   virtual void publishMsg();
+
+  bool getParams(ros::NodeHandle nh);
   bool checkCorrectedData(const pose_estimator::CorrectedData& data);
 
  private:
-  ros::Publisher altitude_heading_behavior_pub;
+  ros::Publisher fixed_rudder_behavior_pub;
 
+  float m_depth;
   float m_altitude;
-  float m_heading;
+  float m_rudder;
   float m_speed_knots;
 
+  std::string m_depth_unit;
   std::string m_altitude_unit;
-  std::string m_heading_unit;
+  std::string m_rudder_unit;
 
   bool m_altitude_ena;
-  bool m_heading_ena;
+  bool m_depth_ena;
+  bool m_rudder_ena;
   bool m_speed_knots_ena;
 
+  float m_depth_tol;
+  float m_rudder_tol;
   float m_altitude_tol;
-  float m_heading_tol;
 };
 
-}  // namespace mission_manager
+}   //  namespace mission_manager
 
-#endif
+#endif  //  MISSION_MANAGER_BEHAVIORS_FIXED_RUDDER_H
