+++ mission_manager/package.xml
@@ -13,6 +13,7 @@
   <build_depend>payload_manager</build_depend>
   <build_depend>pose_estimator</build_depend>
   <build_depend>roscpp</build_depend>
+  <build_depend>roslint</build_depend>
   <build_depend>rospy</build_depend>
   <build_depend>std_msgs</build_depend>
   <build_export_depend>health_monitor</build_export_depend>
